SmartKitchen.AI/requirements.txt
fastapi
uvicorn
python-multipart
pandas
transformers
torch
torchvision
pillow
qwen-vl-utils
accelerate
duckduckgo-search
recipe-scrapers
beautifulsoup4
